pyproject.toml
[tool.poetry]
name = "olx-ads-on-telegram"
version = "0.1.0"
description = "Script that sends new ads from olx to telegram channel"
authors = ["arktnld <arktnld@gmail.com>"]
readme = "README.md"
packages = [{include = "olx_ads_on_telegram"}]

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
telethon = "^1.28.5"
bs4 = "^0.0.1"
apscheduler = "^3.10.1"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
script = "olx_ads_on_telegram.main:start"
